+++ Electronics/Flight firmware/Core/Inc/Drivers/Quaternion.h
@@ -36,14 +36,14 @@
  * Data structure to hold a quaternion.
  */
 typedef struct Quaternion {
-    double w;       /**< Scalar part */
-    double v[3];    /**< Vector part */
+    float w;       /**< Scalar part */
+    float v[3];    /**< Vector part */
 } Quaternion;
 
 /**
  * Sets the given values to the output quaternion.
  */
-void Quaternion_set(double w, double v1, double v2, double v3, Quaternion* output);
+void Quaternion_set(float w, float v1, float v2, float v3, Quaternion* output);
 
 /**
  * Sets quaternion to its identity.
@@ -72,7 +72,7 @@ void Quaternion_fprint(FILE* file, Quaternion* q);
  * @param angle
  *      Rotation angle in radians.
  */
-void Quaternion_fromAxisAngle(double axis[3], double angle, Quaternion* output);
+void Quaternion_fromAxisAngle(float axis[3], float angle, Quaternion* output);
 
 /**
  * Calculates the rotation vector and angle of a quaternion.
@@ -81,48 +81,48 @@ void Quaternion_fromAxisAngle(double axis[3], double angle, Quaternion* output);
  * @return
  *      The rotation angle in radians.
  */
-double Quaternion_toAxisAngle(Quaternion* q, double output[3]);
+float Quaternion_toAxisAngle(Quaternion* q, float output[3]);
 
 /**
  * Set the quaternion to the equivalent of euler angles.
  * @param eulerZYX
  *      Euler angles in ZYX, but stored in array as [x'', y', z].
  */
-void Quaternion_fromEulerZYX(double eulerZYX[3], Quaternion* output);
+void Quaternion_fromEulerZYX(float eulerZYX[3], Quaternion* output);
 
 /**
  * Calculates the euler angles of a quaternion.
  * @param output
  *      Euler angles in ZYX, but stored in array as [x'', y', z].
  */
-void Quaternion_toEulerZYX(Quaternion* q, double output[3]);
+void Quaternion_toEulerZYX(Quaternion* q, float output[3]);
 
 /**
  * Set the quaternion to the equivalent a rotation around the X-axis.
  * @param angle
  *      Rotation angle in radians.
  */
-void Quaternion_fromXRotation(double angle, Quaternion* output);
+void Quaternion_fromXRotation(float angle, Quaternion* output);
 
 /**
  * Set the quaternion to the equivalent a rotation around the Y-axis.
  * @param angle
  *      Rotation angle in radians.
  */
-void Quaternion_fromYRotation(double angle, Quaternion* output);
+void Quaternion_fromYRotation(float angle, Quaternion* output);
 
 /**
  * Set the quaternion to the equivalent a rotation around the Z-axis.
  * @param angle
  *      Rotation angle in radians.
  */
-void Quaternion_fromZRotation(double angle, Quaternion* output);
+void Quaternion_fromZRotation(float angle, Quaternion* output);
 
 /**
  * Calculates the norm of a given quaternion:
  * norm = sqrt(w*w + v1*v1 + v2*v2 + v3*v3)
  */
-double Quaternion_norm(Quaternion* q);
+float Quaternion_norm(Quaternion* q);
 
 /**
  * Normalizes the quaternion.
@@ -146,7 +146,7 @@ void Quaternion_multiply(Quaternion* q1, Quaternion* q2, Quaternion* output);
 /**
  * Applies quaternion rotation to a given vector.
  */
-void Quaternion_rotate(Quaternion* q, double v[3], double output[3]);
+void Quaternion_rotate(Quaternion* q, float v[3], float output[3]);
 
 /**
  * Interpolates between two quaternions.
@@ -154,9 +154,9 @@ void Quaternion_rotate(Quaternion* q, double v[3], double output[3]);
  *      Interpolation between the two quaternions [0, 1].
  *      0 is equal with q1, 1 is equal with q2, 0.5 is the middle between q1 and q2.
  */
-void Quaternion_slerp(Quaternion* q1, Quaternion* q2, double t, Quaternion* output);
+void Quaternion_slerp(Quaternion* q1, Quaternion* q2, float t, Quaternion* output);
 
 
-void Quaternion_fromRate(double omega[3], double dt, Quaternion* output);
+void Quaternion_fromRate(float omega[3], float dt, Quaternion* output);
 
 #endif /* INC_DRIVERS_QUATERNION_H_ */
