+++ Electronics/Flight firmware/Core/Inc/orientation.h
@@ -14,23 +14,23 @@ typedef struct {
     Quaternion orientationQuat;
     Quaternion orientationQuatConj;
     Quaternion incrementalRotation;
-    double gyroVec[3];
-    double accBodyVec[3];
-    double accWorldVec[3];
-    double vertical[3];
+    float gyroVec[3];
+    float accBodyVec[3];
+    float accWorldVec[3];
+    float vertical[3];
     Quaternion gyroQuat;
     Quaternion accQuat;
-    double eulerZYX[3];
+    float eulerZYX[3];
     Quaternion horQuat;
 } Orientation;
 
-void cross_product(double a[3], double b[3], double output[3]);
-void vector_sum(double a[3], double b[3], double output[3]);
+void cross_product(float a[3], float b[3], float output[3]);
+void vector_sum(float a[3], float b[3], float output[3]);
 
 void orientation_init(Orientation *ori);
 
-void orientation_setGyro(Orientation *ori, double gyro[3]);
-void orientation_setAcc(Orientation *ori, double acc[3]);
-void orientation_update(Orientation *ori, double dt);
+void orientation_setGyro(Orientation *ori, float gyro[3]);
+void orientation_setAcc(Orientation *ori, float acc[3]);
+void orientation_update(Orientation *ori, float dt);
 
 #endif /* INC_DRIVERS_ORIENTATION_H_ */
