+++ Electronics/Flight firmware/Core/Src/orientation.c
@@ -9,19 +9,19 @@
 #include "orientation.h"
 #include "constants.h"
 
-void cross_product(double a[3], double b[3], double output[3]) {
+void cross_product(float a[3], float b[3], float output[3]) {
     output[0] = a[1] * b[2] - a[2]*b[1];
     output[1] = a[2] * b[0] - a[0]*b[2];
     output[2] = a[0] * b[1] - a[1]*b[0];
 }
 
-void vector_sum(double a[3], double b[3], double output[3]) {
+void vector_sum(float a[3], float b[3], float output[3]) {
     output[0] = a[0] + b[0];
     output[1] = a[1] + b[1];
     output[2] = a[2] + b[2];
 }
 
-double vector_lenSquared(double a[3]) {
+float vector_lenSquared(float a[3]) {
     return a[0]*a[0] + a[1]*a[1] + a[2]*a[2];
 }
 
@@ -40,29 +40,29 @@ void orientation_init(Orientation *ori) {
     ori->gyroVec[2] = 0;
 }
 
-void orientation_setGyro(Orientation *ori, double gyro[3]) {
+void orientation_setGyro(Orientation *ori, float gyro[3]) {
     ori->gyroVec[0] = gyro[0];
     ori->gyroVec[1] = gyro[1];
     ori->gyroVec[2] = gyro[2];
 }
 
-void orientation_setAcc(Orientation *ori, double acc[3]) {
+void orientation_setAcc(Orientation *ori, float acc[3]) {
     ori->accBodyVec[0] = acc[0];
     ori->accBodyVec[1] = acc[1];
     ori->accBodyVec[2] = acc[2];
 }
 
 // based on https://github.com/daPhoosa/SimpleIMU-6/blob/master/SimpleIMU-6.ino
-void orientation_update(Orientation *ori, double dt) {
+void orientation_update(Orientation *ori, float dt) {
 
     //Quaternion_set(0,ori->gyroVec[0],ori->gyroVec[1],ori->gyroVec[2],&ori->gyroQuat);
     //Quaternion_set(0,ori->accBodyVec[0],ori->accBodyVec[1],ori->accBodyVec[2],&ori->accQuat);
 
     Quaternion_rotate(&ori->orientationQuat, ori->accBodyVec, ori->accWorldVec);
-    double correctionWorld[3];
+    float correctionWorld[3];
     cross_product(ori->accWorldVec, ori->vertical, correctionWorld);
 
-    double correctionBody[3];
+    float correctionBody[3];
     Quaternion_conjugate(&ori->orientationQuat, &ori->orientationQuatConj);
     Quaternion_rotate(&ori->orientationQuatConj, correctionWorld, correctionBody);
 
@@ -70,7 +70,7 @@ void orientation_update(Orientation *ori, double dt) {
     correctionBody[1] = correctionBody[1] * 0.1;
     correctionBody[2] = correctionBody[2] * 0.1;
 
-    double GsSquared = vector_lenSquared(ori->accBodyVec) / (standardGravity * standardGravity);
+    float GsSquared = vector_lenSquared(ori->accBodyVec) / (standardGravity * standardGravity);
     if (GsSquared > 0.81 && GsSquared < 1.21) {
         vector_sum(ori->gyroVec, correctionBody, ori->gyroVec);
     }
